fix: range-check column and scalar type ordinals before narrowing
Review feedback from CodeRabbit on #390: intValue() keeps only the low 32 bits,
so an out-of-range ordinal such as 4294967297 (2^32 + 1) wrapped to 1 and was
accepted as COLUMN_SCALAR — a hole in exactly the case the guard was added for.
The reply's long is now range-checked before it is narrowed.

The same hole existed on the scalar path (ResultSetScalarTypes.getValue via
getValueTypeFromObject), which is the lookup this PR cited as its model, so it
gets the same treatment rather than being left inconsistent. Its exception-driven
bounds check is now an explicit range test.

Regression tests cover the wrapping value and a negative ordinal on both paths.

+++ src/main/java/com/falkordb/impl/resultset/ResultSetImpl.java
@@ -406,7 +406,7 @@ public final class ResultSetImpl implements ResultSet {
      * @return scalar type
      */
     private ResultSetScalarTypes getValueTypeFromObject(Object rawScalarType) {
-        return ResultSetScalarTypes.getValue(((Long) rawScalarType).intValue());
+        return ResultSetScalarTypes.getValue((Long) rawScalarType);
     }
 
     @Override
